refactor(submissions): break submission filters, statistics, and table into components
